Update forgot password to send password reset
The existing forgot your password implementation allows anyone that
knows an email address to reset a users password to a new system
generated one.

This update changes the process so that when someone enters their email
address a change password link is sent to that email address. The link
contains a password reset token and is only valid for 24 hours. The user
can then follow the link to change their password if they wish. The
password is not changed until they submit a new password.